full_pandemic_simulation_tool/requirements.txt

streamlit
crewai
autogen
agentops
composio_core 
composio_openai
together
llama3
ai_ml_api
pandas
matplotlib
plotly
python-dotenv
